pyproject.toml
[project]
name = "hello-langgraph"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "langgraph[langchain] (>=0.4.8,<0.5.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-ollama (>=0.3.3,<0.4.0)",
    "langchain-openai (>=0.2.0,<0.3.0)",
    "jupyterlab (>=4.4.3,<5.0.0)",
    "notebook (>=7.4.3,<8.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)"
]

[tool.poetry]
package-mode = false

packages = [
    {include = "common"},
    {include = "chatbox"},
    {include = "quick-start"},
    {include = "*.py"}
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
